requirements.txt
mysql-connector-python
pywin32
pyodbc
